Adds basic server project to the solution
Adds a basic .NET server project with a simple "Hello, World!" program to the solution.

The Godot project is updated to remove the server directory from compilation.
Also updates the .sln to include the new server project.
Includes a basic .gitignore for the server project.
Adds an .editorconfig to enforce code style.
Adds vcs.xml to track git directory mappings.