Implements basic client-server connection
Adds a basic connection flow to a server via IP address.

A new ConnectMenu scene allows the user to input a server IP and port, and then attempts to connect.
The server now handles incoming connections, storing connected players in a dictionary and logging connection/disconnection events.
The client now polls events and prints to the console that it is trying to connect to a server.
Shared settings have been moved to consts